// Assets-Uncapped/front-portal-develop/src/services/api/loan-operations/models/BillInvoiceDTO.ts
export interface BillInvoiceDTO {
    /**
     * 
     * @type {string}
     * @memberof BillInvoiceDTO
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof BillInvoiceDTO
     */
    organisationId?: string;
    /**
     * 
     * @type {string}
     * @memberof BillInvoiceDTO
     */
    referenceText?: string;
    /**
     * 
     * @type {string}
     * @memberof BillInvoiceDTO
     */
    vendorId?: string;
    /**
     * 
     * @type {string}
     * @memberof BillInvoiceDTO
     */
    paymentId?: string;
    /**
     * 
     * @type {Date}
     * @memberof BillInvoiceDTO
     */
    dateToSendPayment?: Date;
    /**
     * 
     * @type {Date}
     * @memberof BillInvoiceDTO
     */
    creationDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof BillInvoiceDTO
     */
    executionDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof BillInvoiceDTO
     */
    currency?: string;
    /**
     * 
     * @type {number}
     * @memberof BillInvoiceDTO
     */
    amount?: number;
    /**
     * 
     * @type {number}
     * @memberof BillInvoiceDTO
     */
    amountInAgreementCurrency?: number;
    /**
     * 
     * @type {string}
     * @memberof BillInvoiceDTO
     */
    status?: BillInvoiceDTOStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof BillInvoiceDTO
     */
    category?: string;
    /**
     * 
     * @type {string}
     * @memberof BillInvoiceDTO
     */
    paymentStatus?: string;
    /**
     * 
     * @type {string}
     * @memberof BillInvoiceDTO
     */
    paymentMessage?: string;
    /**
     * 
     * @type {string}
     * @memberof BillInvoiceDTO
     */
    rejectReason?: string;
    /**
     * 
     * @type {string}
     * @memberof BillInvoiceDTO
     */
    createdBy?: string;
    /**
     * 
     * @type {string}
     * @memberof BillInvoiceDTO
     */
    decisionBy?: string;
    /**
     * 
     * @type {number}
     * @memberof BillInvoiceDTO
     */
    feeInPercents?: number;
    /**
     * 
     * @type {boolean}
     * @memberof BillInvoiceDTO
     */
    firstStatusNew?: boolean;
    /**
     * 
     * @type {number}
     * @memberof BillInvoiceDTO
     */
    manuallyPaidAmount?: number;
    /**
     * 
     * @type {string}
     * @memberof BillInvoiceDTO
     */
    externalInvoiceId?: string;
    /**
     * 
     * @type {string}
     * @memberof BillInvoiceDTO
     */
    transferType?: string;
    /**
     * 
     * @type {boolean}
     * @memberof BillInvoiceDTO
     */
    awaitingVendorApproval?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof BillInvoiceDTO
     */
    cancelled?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof BillInvoiceDTO
     */
    failed?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof BillInvoiceDTO
     */
    approved?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof BillInvoiceDTO
     */
    paid?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof BillInvoiceDTO
     */
    newInvoice?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof BillInvoiceDTO
     */
    handledManually?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof BillInvoiceDTO
     */
    rejectable?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof BillInvoiceDTO
     */
    cancelable?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof BillInvoiceDTO
     */
    refunded?: boolean;
}


/**
 * @export
 */
export const BillInvoiceDTOStatusEnum = {
    New: 'NEW',
    AwaitingVendorApproval: 'AWAITING_VENDOR_APPROVAL',
    HandleManually: 'HANDLE_MANUALLY',
    Approved: 'APPROVED',
    Rejected: 'REJECTED',
    Refunded: 'REFUNDED',
    ModulrProcessing: 'MODULR_PROCESSING',
    CurrencyCloudProcessing: 'CURRENCY_CLOUD_PROCESSING',
    Cancelled: 'CANCELLED',
    Failed: 'FAILED',
    Paid: 'PAID'
} as const;
export type BillInvoiceDTOStatusEnum = typeof BillInvoiceDTOStatusEnum[keyof typeof BillInvoiceDTOStatusEnum];