// Assets-Uncapped/front-portal-develop/src/services/api/loan-operations/models/BillPaymentFee.ts
/* tslint:disable */
/* eslint-disable */
/**
 * api-loan-operations
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: Unknown version
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BillPaymentFee
 */
export interface BillPaymentFee {
    /**
     * 
     * @type {number}
     * @memberof BillPaymentFee
     */
    marketing?: number;
    /**
     * 
     * @type {number}
     * @memberof BillPaymentFee
     */
    inventory?: number;
    /**
     * 
     * @type {number}
     * @memberof BillPaymentFee
     */
    other?: number;
    /**
     * 
     * @type {number}
     * @memberof BillPaymentFee
     */
    exchange?: number;
}

/**
 * Check if a given object implements the BillPaymentFee interface.
 */
export function instanceOfBillPaymentFee(value: object): value is BillPaymentFee {
    return true;
}

export function BillPaymentFeeFromJSON(json: any): BillPaymentFee {
    return BillPaymentFeeFromJSONTyped(json, false);
}

export function BillPaymentFeeFromJSONTyped(json: any, ignoreDiscriminator: boolean): BillPaymentFee {
    if (json == null) {
        return json;
    }
    return {
        
        'marketing': json['marketing'] == null ? undefined : json['marketing'],
        'inventory': json['inventory'] == null ? undefined : json['inventory'],
        'other': json['other'] == null ? undefined : json['other'],
        'exchange': json['exchange'] == null ? undefined : json['exchange'],
    };
}

export function BillPaymentFeeToJSON(json: any): BillPaymentFee {
    return BillPaymentFeeToJSONTyped(json, false);
}

export function BillPaymentFeeToJSONTyped(value?: BillPaymentFee | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'marketing': value['marketing'],
        'inventory': value['inventory'],
        'other': value['other'],
        'exchange': value['exchange'],
    };
}